+++ src/rapids_singlecell/preprocessing/_pca.py
@@ -5,14 +5,15 @@ from typing import TYPE_CHECKING
 
 import cupy as cp
 import numpy as np
-from cuml.decomposition import PCA, TruncatedSVD
+from cuml.decomposition import TruncatedSVD
 from cuml.internals.input_utils import sparse_scipy_to_cp
-from cupyx.scipy.sparse import csr_matrix, isspmatrix_csr
+from cupyx.scipy.sparse import csr_matrix
 from cupyx.scipy.sparse import issparse as cpissparse
 from scanpy._utils import Empty, _empty
 from scanpy.preprocessing._pca import _handle_mask_var
 from scipy.sparse import issparse
 
+from rapids_singlecell._compat import DaskArray, DaskClient
 from rapids_singlecell.get import _get_obs_rep
 
 if TYPE_CHECKING:
@@ -34,6 +35,7 @@ def pca(
     copy: bool = False,
     chunked: bool = False,
     chunk_size: int = None,
+    client: DaskClient | None = None,
 ) -> None | AnnData:
     """
     Performs PCA using the cuml decomposition function.
@@ -88,6 +90,9 @@ def pca(
             Number of observations to include in each chunk. \
             Required if `chunked=True` was passed.
 
+        client
+            Dask client to use for computation. If `None`, the default client is used. Only used if `X` is a Dask array.
+
     Returns
     -------
         adds fields to `adata`:
@@ -126,50 +131,70 @@ def pca(
             n_comps = min_dim - 1
         else:
             n_comps = 50
+    if isinstance(X, DaskArray):
+        if chunked:
+            raise ValueError(
+                "Dask arrays are not supported for chunked PCA computation."
+            )
+        if isinstance(X._meta, cp.ndarray):
+            from cuml.dask.decomposition import PCA
 
-    if chunked:
-        from cuml.decomposition import IncrementalPCA
+            if svd_solver == "auto":
+                svd_solver = "jacobi"
+            pca_func = PCA(n_components=n_comps, svd_solver=svd_solver)
+            X_pca = pca_func.fit_transform(X)
+        elif isinstance(X._meta, csr_matrix):
+            from cuml.dask.decomposition import PCA_sparse
 
-        X_pca = np.zeros((X.shape[0], n_comps), X.dtype)
+            pca_func = PCA_sparse(n_components=n_comps)
+            X_pca = pca_func.fit_transform(X)
 
-        pca_func = IncrementalPCA(
-            n_components=n_comps, output_type="numpy", batch_size=chunk_size
-        )
-        pca_func.fit(X)
-
-        n_batches = math.ceil(X.shape[0] / chunk_size)
-        for batch in range(n_batches):
-            start_idx = batch * chunk_size
-            stop_idx = min(batch * chunk_size + chunk_size, X.shape[0])
-            chunk = X[start_idx:stop_idx, :]
-            if issparse(chunk) or cpissparse(chunk):
-                chunk = chunk.toarray()
-            X_pca[start_idx:stop_idx] = pca_func.transform(chunk)
     else:
-        if zero_center:
-            if cpissparse(X) or issparse(X):
-                if issparse(X):
-                    X = sparse_scipy_to_cp(X, dtype=X.dtype)
-                    X = csr_matrix(X)
-                pca_func = PCA_sparse(n_components=n_comps)
-                X_pca = pca_func.fit_transform(X)
-            else:
-                pca_func = PCA(
+        if chunked:
+            from cuml.decomposition import IncrementalPCA
+
+            X_pca = np.zeros((X.shape[0], n_comps), X.dtype)
+
+            pca_func = IncrementalPCA(
+                n_components=n_comps, output_type="numpy", batch_size=chunk_size
+            )
+            pca_func.fit(X)
+
+            n_batches = math.ceil(X.shape[0] / chunk_size)
+            for batch in range(n_batches):
+                start_idx = batch * chunk_size
+                stop_idx = min(batch * chunk_size + chunk_size, X.shape[0])
+                chunk = X[start_idx:stop_idx, :]
+                if issparse(chunk) or cpissparse(chunk):
+                    chunk = chunk.toarray()
+                X_pca[start_idx:stop_idx] = pca_func.transform(chunk)
+        else:
+            if zero_center:
+                if cpissparse(X) or issparse(X):
+                    if issparse(X):
+                        X = sparse_scipy_to_cp(X, dtype=X.dtype)
+                        X = csr_matrix(X)
+                    from ._sparse_pca._sparse_pca import PCA_sparse
+
+                    pca_func = PCA_sparse(n_components=n_comps)
+                    X_pca = pca_func.fit_transform(X)
+                else:
+                    pca_func = PCA(
+                        n_components=n_comps,
+                        svd_solver=svd_solver,
+                        random_state=random_state,
+                        output_type="numpy",
+                    )
+                    X_pca = pca_func.fit_transform(X)
+
+            elif not zero_center:
+                pca_func = TruncatedSVD(
                     n_components=n_comps,
-                    svd_solver=svd_solver,
                     random_state=random_state,
+                    algorithm=svd_solver,
                     output_type="numpy",
                 )
                 X_pca = pca_func.fit_transform(X)
-
-        elif not zero_center:
-            pca_func = TruncatedSVD(
-                n_components=n_comps,
-                random_state=random_state,
-                algorithm=svd_solver,
-                output_type="numpy",
-            )
-            X_pca = pca_func.fit_transform(X)
 
     if X_pca.dtype.descr != np.dtype(dtype).descr:
         X_pca = X_pca.astype(dtype)
@@ -194,187 +219,3 @@ def pca(
         adata.varm["PCs"] = pca_func.components_.T
     if copy:
         return adata
-
-
-class PCA_sparse:
-    def __init__(self, n_components) -> None:
-        self.n_components = n_components
-
-    def fit(self, x):
-        if self.n_components is None:
-            n_rows = x.shape[0]
-            n_cols = x.shape[1]
-            self.n_components_ = min(n_rows, n_cols)
-        else:
-            self.n_components_ = self.n_components
-
-        if not isspmatrix_csr(x):
-            x = x.tocsr()
-        _check_matrix_for_zero_genes(x)
-        self.n_samples_ = x.shape[0]
-        self.n_features_in_ = x.shape[1] if x.ndim == 2 else 1
-        self.dtype = x.data.dtype
-
-        covariance, self.mean_, _ = _cov_sparse(x=x, return_mean=True)
-
-        self.explained_variance_, self.components_ = cp.linalg.eigh(
-            covariance, UPLO="U"
-        )
-
-        # NOTE: We reverse the eigen vector and eigen values here
-        # because cupy provides them in ascending order. Make a copy otherwise
-        # it is not C_CONTIGUOUS anymore and would error when converting to
-        # CumlArray
-        self.explained_variance_ = self.explained_variance_[::-1]
-
-        self.components_ = cp.flip(self.components_, axis=1)
-
-        self.components_ = self.components_.T[: self.n_components_, :]
-
-        self.explained_variance_ratio_ = self.explained_variance_ / cp.sum(
-            self.explained_variance_
-        )
-
-        self.explained_variance_ = self.explained_variance_[: self.n_components_]
-
-        self.explained_variance_ratio_ = self.explained_variance_ratio_[
-            : self.n_components_
-        ]
-
-        return self
-
-    def transform(self, X):
-        X = X - self.mean_
-        X_transformed = X.dot(self.components_.T)
-        self.components_ = self.components_.get()
-        self.explained_variance_ = self.explained_variance_.get()
-        self.explained_variance_ratio_ = self.explained_variance_ratio_.get()
-        return X_transformed.get()
-
-    def fit_transform(self, X, y=None):
-        return self.fit(X).transform(X)
-
-
-def _cov_sparse(x, return_gram=False, return_mean=False):
-    """
-    Computes the mean and the covariance of matrix X of
-    the form Cov(X, X) = E(XX) - E(X)E(X)
-
-    This is a temporary fix for
-    cuml issue #5475 and cupy issue #7699,
-    where the operation `x.T.dot(x)` did not work for
-    larger sparse matrices.
-
-    Parameters
-    ----------
-
-    x : cupyx.scipy.sparse of size (m, n)
-    return_gram : boolean (default = False)
-        If True, gram matrix of the form (1 / n) * X.T.dot(X)
-        will be returned.
-        When True, a copy will be created
-        to store the results of the covariance.
-        When False, the local gram matrix result
-        will be overwritten
-    return_mean: boolean (default = False)
-        If True, the Maximum Likelihood Estimate used to
-        calculate the mean of X and X will be returned,
-        of the form (1 / n) * mean(X) and (1 / n) * mean(X)
-
-    Returns
-    -------
-
-    result : cov(X, X) when return_gram and return_mean are False
-            cov(X, X), gram(X, X) when return_gram is True,
-            return_mean is False
-            cov(X, X), mean(X), mean(X) when return_gram is False,
-            return_mean is True
-            cov(X, X), gram(X, X), mean(X), mean(X)
-            when return_gram is True and return_mean is True
-    """
-
-    from ._kernels._pca_sparse_kernel import (
-        _copy_kernel,
-        _cov_kernel,
-        _gramm_kernel_csr,
-    )
-
-    gram_matrix = cp.zeros((x.shape[1], x.shape[1]), dtype=x.data.dtype)
-
-    block = (128,)
-    grid = (x.shape[0],)
-    compute_mean_cov = _gramm_kernel_csr(x.data.dtype)
-    compute_mean_cov(
-        grid,
-        block,
-        (
-            x.indptr,
-            x.indices,
-            x.data,
-            x.shape[0],
-            x.shape[1],
-            gram_matrix,
-        ),
-    )
-
-    copy_gram = _copy_kernel(x.data.dtype)
-    block = (32, 32)
-    grid = (math.ceil(x.shape[1] / block[0]), math.ceil(x.shape[1] / block[1]))
-    copy_gram(
-        grid,
-        block,
-        (gram_matrix, x.shape[1]),
-    )
-
-    mean_x = x.sum(axis=0) * (1 / x.shape[0])
-    gram_matrix *= 1 / x.shape[0]
-
-    if return_gram:
-        cov_result = cp.zeros(
-            (gram_matrix.shape[0], gram_matrix.shape[0]),
-            dtype=gram_matrix.dtype,
-        )
-    else:
-        cov_result = gram_matrix
-
-    compute_cov = _cov_kernel(x.dtype)
-
-    block_size = (32, 32)
-    grid_size = (math.ceil(gram_matrix.shape[0] / 8),) * 2
-    compute_cov(
-        grid_size,
-        block_size,
-        (cov_result, gram_matrix, mean_x, mean_x, gram_matrix.shape[0]),
-    )
-
-    if not return_gram and not return_mean:
-        return cov_result
-    elif return_gram and not return_mean:
-        return cov_result, gram_matrix
-    elif not return_gram and return_mean:
-        return cov_result, mean_x, mean_x
-    elif return_gram and return_mean:
-        return cov_result, gram_matrix, mean_x, mean_x
-
-
-def _check_matrix_for_zero_genes(X):
-    gene_ex = cp.zeros(X.shape[1], dtype=cp.int32)
-
-    from ._kernels._pca_sparse_kernel import _zero_genes_kernel
-
-    block = (32,)
-    grid = (int(math.ceil(X.nnz / block[0])),)
-    _zero_genes_kernel(
-        grid,
-        block,
-        (
-            X.indices,
-            gene_ex,
-            X.nnz,
-        ),
-    )
-    if cp.any(gene_ex == 0):
-        raise ValueError(
-            "There are genes with zero expression. "
-            "Please remove them before running PCA."
-        )
